go-get-button (component): create this component
