+++ src/lib/components/molecules/go-get-button/index.ts
@@ -1,0 +1,1 @@
+export { GoGetButton } from "./go-get-button";
